refactor: Data structures refactored to improve consistency
