+++ C#/JsonSerializationTests/bin/Debug/netcoreapp3.1/nunit_random_seed.tmp
@@ -1,1 +1,1 @@
-1181687461
+202949533
